Revert "Merge branch 'Branch_Bryan'"
This reverts commit a33bc72d4f6116d69a1daee017cf46e23913fc70, reversing
changes made to 6a3e08ffe6771657ac8b8b37397347de582e49d8.

+++ liste.h
@@ -198,7 +198,7 @@ T Liste<T>::remove(size_t index){
 template <class T>
 T Liste<T>::get(size_t index){      // Element an der Stelle index zurückgeben
   static element<T> *lastTemp = head;
-  static int lastIndex = 0;
+  static size_t lastIndex = 0;
   struct element<T> *temp;
   size_t i;
 
@@ -247,7 +247,6 @@ size_t Liste<T>::number_Elements(){
     return i;
   }
 }
-
 
 
 #endif
